+++ Realm/ObjectServerTests/RLMSyncTestCase.m
@@ -74,12 +74,10 @@ static NSURL *syncDirectoryForChildProcess() {
     return [NSURL URLWithString:@"http://127.0.0.1:9080"];
 }
 
-+ (RLMSyncCredential *)basicCredentialWithName:(NSString *)name createAccount:(BOOL)createAccount {
++ (RLMSyncCredential *)basicCredentialWithName:(NSString *)name register:(BOOL)shouldRegister {
     return [RLMSyncCredential credentialWithUsername:name
                                             password:@"a"
-                                             actions:(createAccount
-                                                      ? RLMAuthenticationActionsCreateAccount
-                                                      : RLMAuthenticationActionsUseExistingAccount)];
+                                            register:shouldRegister];
 }
 
 - (void)addSyncObjectsToRealm:(RLMRealm *)realm descriptions:(NSArray<NSString *> *)descriptions {
@@ -137,16 +135,16 @@ static NSURL *syncDirectoryForChildProcess() {
     NSString *process = self.isParent ? @"parent" : @"child";
     __block RLMSyncUser *theUser = nil;
     XCTestExpectation *expectation = [self expectationWithDescription:@"Should log in the user properly"];
-    [RLMSyncUser authenticateWithCredential:credential
-                              authServerURL:url
-                               onCompletion:^(RLMSyncUser *user, NSError *error) {
-                                   XCTAssertNil(error,
-                                                @"Error when trying to log in a user: %@ (process: %@)",
-                                                error, process);
-                                   XCTAssertNotNil(user);
-                                   theUser = user;
-                                   [expectation fulfill];
-                               }];
+    [RLMSyncUser logInWithCredential:credential
+                       authServerURL:url
+                        onCompletion:^(RLMSyncUser *user, NSError *error) {
+                            XCTAssertNil(error,
+                                         @"Error when trying to log in a user: %@ (process: %@)",
+                                         error, process);
+                            XCTAssertNotNil(user);
+                            theUser = user;
+                            [expectation fulfill];
+                        }];
     [self waitForExpectationsWithTimeout:4.0 handler:nil];
     XCTAssertTrue(theUser.state == RLMSyncUserStateActive,
                   @"User should have been valid, but wasn't. (process: %@)", process);